fixed white space changes, moved ConeCastUtility to the right namespace
